lingbin-temp-update: temp-commit-at-20240106
lingbin-temp-update: temp-commit-at-20240204

lingbin-temp-update: temp-commit-at-20240215

lingbin-temp-update: temp-commit-at-20240215-2

lingbin-temp-update: temp-commit-at-20240216

lingbin-temp-update: temp-commit-at-20240217

lingbin-temp-update: temp-commit-at-20240224

lingbin-temp-update: temp-commit-at-20240226

lingbin-temp-update: temp-commit-at-20240308

lingbin-temp-update: temp-commit-at-20240316

lingbin-temp-update: temp-commit-at-20240331

lingbin-temp-update: temp-commit-at-20240417

lingbin-temp-update: temp-commit-at-20240420

lingbin-temp-update: temp-commit-at-20240426

lingbin-temp-update: temp-commit-at-20240514

lingbin-temp-update: temp-commit-at-20240516

lingbin-temp-update: temp-commit-at-20240623

lingbin-temp-update: temp-commit-at-20240627

lingbin-temp: fix Driver.h

lingbin-temp: fix VectorHasher.cpp for 20240514

lingbin-temp: fix VectorHasher.cpp for 20240630

lingbin-temp: fix VectorHasher.cpp for 20240702

Fix PeriodicStatsReporter.h on MocOS(clang)

lingbin-temp: temp-commit-at-20240712

lingbin-temp: temp-commit-at-20240714: MmapArea

lingbin-temp: temp-commit-at-20240714: MemoryAllocator

lingbin-temp: temp-commit-at-20240714: MemoryAllocator

lingbin-temp: temp-commit-at-20240717: Driver

lingbin-temp: temp-commit-at-20240720: Task

lingbin-temp: temp-commit-at-20240806: exec-ExchangeClient

lingbin-temp: temp-commit-at-20240806: type-Filter

lingbin-temp: temp-commit-at-20240806: connector

lingbin-temp: temp-commit-at-20240810: connector

lingbin-temp: temp-commit-at-20240823: Vector

lingbin-temp: temp-commit-at-20240825: Connector

lingbin-temp: temp-commit-at-20240826: HashProbe-remove copy

+++ velox/common/memory/tests/StreamArenaTest.cpp
@@ -110,9 +110,9 @@ TEST_F(StreamArenaTest, newRange) {
         testData.expectedContiguousAllocationSize +
             testData.expectedNonContiguousAllocationSize,
         arena->size());
-    const int allocateBytes =
+    const uint64_t allocateBytes =
         AllocationTraits::pageBytes(mmapAllocator_->numAllocated());
-    const int contiguousBytes =
+    const uint64_t contiguousBytes =
         AllocationTraits::pageBytes(mmapAllocator_->numExternalMapped());
     ASSERT_EQ(
         testData.expectedNonContiguousAllocationSize,
@@ -127,13 +127,13 @@ TEST_F(StreamArenaTest, randomRange) {
   ByteRange range;
   for (int i = 0; i < numRanges; ++i) {
     if (folly::Random::oneIn(3)) {
-      const int requestSize =
+      const uint64_t requestSize =
           AllocationTraits::pageBytes(pool_->largestSizeClass()) +
           (folly::Random::rand32() % (4 << 20));
       arena->newRange(requestSize, nullptr, &range);
       ASSERT_EQ(AllocationTraits::roundUpPageBytes(requestSize), range.size);
     } else {
-      const int requestSize =
+      const uint64_t requestSize =
           1 + folly::Random::rand32() % pool_->largestSizeClass();
       arena->newRange(requestSize, nullptr, &range);
       ASSERT_LE(range.size, AllocationTraits::roundUpPageBytes(requestSize));
